requirements.txt
streamlit
python-dotenv
google-generativeai
easyocr
tqdm
pdf2image
